requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
filelock==3.0.12
Flask==1.1.2
future==0.18.2
google-api-core==1.22.2
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.27.2
google-cloud-core==1.4.1
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
nltk==3.5
numpy==1.19.2
oauthlib==3.1.0
packaging==20.4
pandas==1.1.2
pandas-gbq==0.13.2
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.2
sentence-transformers==0.3.6
sentencepiece==0.1.91
six==1.15.0
threadpoolctl==2.1.0
tokenizers==0.8.1rc2
torch==1.6.0+cpu
torchvision==0.7.0+cpu
tqdm==4.49.0
transformers==3.1.0
urllib3==1.25.10
Werkzeug==1.0.1
